aetherium-backend/requirements.txt
aiosqlite
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
chromadb 
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
ollama
sniffio==1.3.1
starlette==0.47.3
sqlalchemy[asyncio]
sentence-transformers
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1

